// dashboard/src/client/models/worker-package-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const WorkerPackageStatus = {
    Installed: 'installed',
    Installing: 'installing',
    Uninstalling: 'uninstalling',
    Removed: 'removed',
    Failed: 'failed',
    Missing: 'missing',
    Unknown: 'unknown'
} as const;

export type WorkerPackageStatus = typeof WorkerPackageStatus[keyof typeof WorkerPackageStatus];